backend_incapacidades_cloud_mix_v1/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
# Supabase Python client
supabase==2.6.0
